// src/store/types/Commintment.ts
export type Commintment = {
    actionPlanId: number;
    commitmentId: number;
    commitmentStatusDescription: string;
    commitmentStatusId: number;
    description: string;
    isActive: boolean;
    CreationDate: Date; // Ajusta según el tipo real de CreationDate
    ModificationDate: Date; // Ajusta según el tipo real de ModificationDate
};